web_streamlit/app/requirements.txt
pillow-heif
Pillow
numpy
# img
pillow

# web
streamlit
requests
python-dotenv

opencv-python-headless==4.6.0.66
datetime

# Live video
streamlit-webrtc
